+++ src/bedrock/symbols/windows.h
@@ -11,7 +11,7 @@
 
 namespace endstone::runtime {
 
-static constexpr std::array<std::pair<std::string_view, std::size_t>, 73> symbols = {{
+static constexpr std::array<std::pair<std::string_view, std::size_t>, 74> symbols = {{
     {"BlockState::StateListNode::mHead", 211294960},
     {"Enchant::mEnchants", 211141912},
     {"MobEffect::mMobEffects", 211424880},
@@ -104,6 +104,7 @@ static constexpr std::array<std::pair<std::string_view, std::size_t>, 73> symbol
     {"?setRespawnPosition@Player@@QEAAXAEBVBlockPos@@UDimensionType@@@Z", 2307600},
     {"?setSpawnBlockRespawnPosition@Player@@QEAA_NAEBVBlockPos@@UDimensionType@@@Z", 2306752},
     {"?startSleepInBed@Player@@UEAA?AW4BedSleepingResult@@AEBVBlockPos@@_NM@Z", 2182176},
+    {"?stopSleepInBed@Player@@UEAAX_N0@Z", 2187632},
     {"?take@Player@@QEAA_NAEAVActor@@HH@Z", 2251552},
     {"?teleportTo@Player@@UEAAXAEBVVec3@@_NHH1@Z", 2176288},
     // PushableByEntityUtility